refactor: ApiErrorResponse 경로 수정
