+++ src/apis/festivals/postFestivalsPick.ts
@@ -1,8 +1,7 @@
 import { apiInstance } from '@/apis/apiInstance';
 import API_ENDPOINTS from '@/constants/apiEndpoints';
 import type { Festival } from '@/types/FestivalType';
-import type { ApiResponseList } from '@/apis/apiResponse';
-import type { ApiErrorResponse } from '@/apis/apiInstance';
+import type { ApiResponseList, ApiErrorResponse } from '@/apis/apiResponse';
 import type { AxiosResponse } from 'axios';
 import type { PickStyleId } from '@/constants/pickStyles';
