+++ src/apis/review/getSingleReview.ts
@@ -1,9 +1,8 @@
 import { apiInstance } from '@/apis/apiInstance';
 import API_ENDPOINTS from '@/constants/apiEndpoints';
-import type { ApiErrorResponse } from '@/apis/apiInstance';
+import type { ApiResponseItem, ApiErrorResponse } from '@/apis/apiResponse';
 import type { AxiosResponse } from 'axios';
-import type { Review } from './getReview';
-import type { ApiResponseItem } from '../apiResponse';
+import type { Review } from '@/apis/review/getReview';
 import { generatePath } from 'react-router-dom';
 
 export const getSingleReview = async (
