+++ src/apis/user/getUserInfo.ts
@@ -1,7 +1,7 @@
 import { apiInstance } from '@/apis/apiInstance';
 import API_ENDPOINTS from '@/constants/apiEndpoints';
 import type { UserInfoResponse } from '@/types/UserType';
-import type { ApiErrorResponse } from '@/apis/apiInstance';
+import type { ApiErrorResponse } from '@/apis/apiResponse';
 import type { AxiosResponse } from 'axios';
 
 export const getUserInfo = async (): Promise<AxiosResponse<UserInfoResponse, ApiErrorResponse>> => {
