+++ src/apis/user/getUserRole.ts
@@ -1,5 +1,5 @@
 import { apiInstance } from '@/apis/apiInstance';
-import type { ApiErrorResponse } from '@/apis/apiInstance';
+import type { ApiErrorResponse } from '@/apis/apiResponse';
 import type { AxiosResponse } from 'axios';
 import API_ENDPOINTS from '@/constants/apiEndpoints';
